pyproject.toml
[project]
name = "portal-datastore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "python-irodsclient>=3.1.1",
]
